Fix -Wextra warnings r20: sign-compare, missing-designated-initializers in distributed/, firmware/
- pcie_core_writer.cpp: add .sdesc_path={} to ClusterOptions
- thread_pool.cpp: fix size_t/int comparison, uint32_t loop var
- multi_producer_single_consumer_queue.hpp: uint32_t loop var
- distributed_host_buffer.cpp: add .buffer={} to Shard designated inits
- mesh_trace.cpp: add .sharding_args={} to DeviceLocalBufferConfig
- mesh_workload.cpp: add .sharding_args={}, fix size_t loop var
- mesh_buffer.cpp: add .bottom_up=nullopt / .sharding_args={} to DeviceLocalBufferConfig
- mesh_device_view.cpp: size_t loop vars for col/row
- mesh_socket_utils.cpp: cast int to size_t in TT_FATAL comparisons
- mesh_command_queue_base.cpp: uint32_t loop vars
- fd_mesh_command_queue.cpp: uint32_t/size_t loop vars, ptrdiff_t cast
- mesh_device.cpp: size_t loop var
- metal_context.cpp: fix int/size_t comparisons with MAX_CONTEXT_COUNT
- system_mesh.cpp: size_t/uint32_t loop vars
- dispatch_kernel_initializer.cpp: cast uint32_t device id to int in lambdas
- command_queue_initializer.cpp: same
- profiler_initializer.cpp: same
- risc_firmware_initializer.cpp: size_t channel, cast num_build_states

+++ tt_metal/distributed/fd_mesh_command_queue.cpp
@@ -863,7 +863,7 @@ void FDMeshCommandQueue::copy_buffer_data_to_user_space(MeshBufferReadDescriptor
         uint16_t channel =
             tt::tt_metal::MetalContext::instance(cid).get_cluster().get_assigned_channel_for_device(device->id());
 
-        for (int i = 0; i < num_reads; i++) {
+        for (uint32_t i = 0; i < num_reads; i++) {
             buffer_dispatch::copy_completion_queue_data_into_user_space(
                 std::get<ReadBufferDescriptor>(*(read_descriptor_queue.pop())),
                 mmio_device_id,
@@ -1091,7 +1091,7 @@ static VecIt remove_by_index(VecIt begin, VecIt end, IndexIt index_begin, IndexI
         if (index_begin == index_end) {
             return false;
         }
-        if (*index_begin == (&value - &*begin)) {
+        if (*index_begin == static_cast<size_t>(&value - &*begin)) {
             ++index_begin;
             return true;
         }
@@ -1161,7 +1161,7 @@ void FDMeshCommandQueue::record_end() {
     command_sequence.add_prefetch_exec_buf_end();
 
     exec_buf_end.reserve(command_sequence.size_bytes() / sizeof(uint32_t));
-    for (int i = 0; i < command_sequence.size_bytes() / sizeof(uint32_t); i++) {
+    for (size_t i = 0; i < command_sequence.size_bytes() / sizeof(uint32_t); i++) {
         exec_buf_end.push_back(static_cast<uint32_t*>(command_sequence.data())[i]);
     }
     size_t max_trace_size = 0;
